prot2/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0

torch==2.0.1
torchvision==0.15.2

numpy
opencv-python
Pillow
tqdm
requests
